@@ -2,64 +2,71 @@
   "name": "inn-calculator",
   "version": "0.0.0",
   "license": "MIT",
+  "private": true,
   "scripts": {
     "debug:inn-calculator": "nx serve",
     "start": "nx serve",
     "build": "nx build",
-    "test": "nx test"
+    "test": "nx test",
+    "build_prod":"nx build --prod --base-href=/inn-calculator/"
   },
-  "private": true,
+  "publishConfig": {
+    "registry": "https://gitlab.company.com/api/v4/projects/44237280/packages/npm/"
+  },
   "dependencies": {
-    "@angular/animations": "~15.2.0",
-    "@angular/common": "~15.2.0",
-    "@angular/compiler": "~15.2.0",
-    "@angular/core": "~15.2.0",
-    "@angular/forms": "~15.2.0",
-    "@angular/platform-browser": "~15.2.0",
-    "@angular/platform-browser-dynamic": "~15.2.0",
-    "@angular/router": "~15.2.0",
-    "@nrwl/angular": "15.8.9",
+    "@angular/animations": "16.1.8",
+    "@angular/common": "16.1.8",
+    "@angular/compiler": "16.1.8",
+    "@angular/core": "16.1.8",
+    "@angular/fire": "^7.6.1",
+    "@angular/forms": "16.1.8",
+    "@angular/platform-browser": "16.1.8",
+    "@angular/platform-browser-dynamic": "16.1.8",
+    "@angular/router": "16.1.8",
+    "@nrwl/angular": "16.6.0",
+    "@nrwl/nx-cloud": "16.2.0",
     "@yarnpkg/lockfile": "^1.1.0",
     "@yarnpkg/parsers": "^3.0.0-rc.40",
     "daisyui": "^2.51.4",
+    "parcel": "^2.9.3",
     "rxjs": "~7.8.0",
     "tailwindcss": "^3.2.7",
     "tslib": "^2.3.0",
-    "zone.js": "~0.12.0"
+    "zone.js": "0.13.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~15.2.0",
-    "@angular-devkit/core": "~15.2.0",
-    "@angular-devkit/schematics": "~15.2.0",
-    "@angular-eslint/eslint-plugin": "~15.0.0",
-    "@angular-eslint/eslint-plugin-template": "~15.0.0",
-    "@angular-eslint/template-parser": "~15.0.0",
+    "@angular-devkit/build-angular": "16.1.7",
+    "@angular-devkit/core": "16.1.7",
+    "@angular-devkit/schematics": "16.1.7",
+    "@angular-eslint/eslint-plugin": "16.0.3",
+    "@angular-eslint/eslint-plugin-template": "16.0.3",
+    "@angular-eslint/template-parser": "16.0.3",
     "@angular/cli": "~15.2.0",
-    "@angular/compiler-cli": "~15.2.0",
-    "@angular/language-service": "~15.2.0",
-    "@nrwl/cypress": "15.8.9",
-    "@nrwl/eslint-plugin-nx": "15.8.9",
-    "@nrwl/jest": "15.8.9",
-    "@nrwl/js": "15.8.9",
-    "@nrwl/linter": "15.8.9",
-    "@nrwl/workspace": "15.8.9",
-    "@schematics/angular": "~15.2.0",
+    "@angular/compiler-cli": "16.1.8",
+    "@angular/language-service": "16.1.8",
+    "@nrwl/cypress": "16.6.0",
+    "@nrwl/eslint-plugin-nx": "16.6.0",
+    "@nrwl/jest": "16.6.0",
+    "@nrwl/js": "16.6.0",
+    "@nrwl/linter": "16.6.0",
+    "@nrwl/workspace": "16.6.0",
+    "@schematics/angular": "16.1.7",
     "@types/jest": "^29.5.0",
     "@types/mocha": "^10.0.1",
     "@types/node": "16.11.7",
-    "@typescript-eslint/eslint-plugin": "^5.36.1",
-    "@typescript-eslint/parser": "^5.36.1",
-    "cypress": "^12.2.0",
+    "@typescript-eslint/eslint-plugin": "5.62.0",
+    "@typescript-eslint/parser": "5.62.0",
+    "cypress": "12.17.3",
     "eslint": "~8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
     "jest": "^29.4.1",
     "jest-environment-jsdom": "^29.4.1",
-    "jest-preset-angular": "~13.0.0",
-    "nx": "15.8.9",
+    "jest-preset-angular": "13.1.1",
+    "nx": "16.6.0",
     "prettier": "^2.6.2",
-    "ts-jest": "^29.0.5",
+    "ts-jest": "29.1.1",
     "ts-node": "10.9.1",
-    "typescript": "~4.9.5"
+    "typescript": "5.1.6"
   }
 }
