@@ -1,19 +1,19 @@
-import { inject, NgModule } from '@angular/core';
+import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ProductListViewComponent } from './components/product-list-view/product-list-view.component';
 
 import { RouterModule, Routes } from '@angular/router';
 import { CalculatorViewComponent } from './components/calculator-view/calculator-view.component';
 import { BillViewComponent } from './components/bill-view/bill-view.component';
-import { ProductResolver } from './resolvers/product.resolver';
-import { ProductCardComponent } from './components/product-card/product-card.component';
+import { CalculatorTabsComponent } from './components/calculator-view/tabs/calculator-tabs/calculator-tabs.component';
+import { ProductListAreaComponent } from './components/product-list-area/product-list-area.component';
+import { ProductListComponent } from './components/product-list/product-list.component';
 
 export const APP_ROUTES: Routes = [
   {
     path: '',
-    component: CalculatorViewComponent,
-    resolve: { product: () => inject(ProductResolver).resolve() },
-  },
+    component: CalculatorViewComponent
+  }
 ];
 
 @NgModule({
@@ -21,11 +21,14 @@ export const APP_ROUTES: Routes = [
     ProductListViewComponent,
     CalculatorViewComponent,
     BillViewComponent,
-    ProductCardComponent,
+    CalculatorTabsComponent,
+    ProductListAreaComponent,
+    ProductListComponent
   ],
+  exports: [],
   imports: [
     CommonModule,
     RouterModule.forRoot(APP_ROUTES, { paramsInheritanceStrategy: 'always' }),
   ],
 })
-export class CalculatorModule {}
+export class CalculatorModule { }
